created register, login

+++ controllers/contacts.js
@@ -4,7 +4,18 @@ const { ctrlWrapper, HttpError } = require("../helpers");
 const schemas = require("../schemas/contactsSchema");
 
 const getList = async (req, res) => {
-  const result = await Contact.find();
+  const { _id: owner } = req.user;
+  const { page = 1, limit = 10, favorite } = req.query;
+  const skip = (page - 1) * limit;
+
+  const result = await Contact.find(
+    { owner, ...(favorite !== undefined && { favorite }) },
+    "-createdAt -updatedAt",
+    {
+      skip,
+      limit,
+    }
+  ).populate("owner", "subscription email");
   res.status(200).json(result);
 };
 
@@ -13,13 +24,15 @@ const add = async (req, res) => {
   if (error) {
     throw HttpError(400, error.message);
   }
-  const result = await Contact.create(req.body);
+  const { _id: owner } = req.user;
+  const result = await Contact.create({ ...req.body, owner });
   res.status(201).json(result);
 };
 
 const getById = async (req, res) => {
   const { id } = req.params;
-  const result = await Contact.findById(id);
+  const { _id: owner } = req.user;
+  const result = await Contact.findOne({ _id: id, owner });
   if (!result) {
     throw HttpError(404, `Contact with a such id=${id} not found`);
   }
@@ -28,7 +41,8 @@ const getById = async (req, res) => {
 
 const removeById = async (req, res) => {
   const { id } = req.params;
-  const result = await Contact.findByIdAndRemove(id);
+  const { _id: owner } = req.user;
+  const result = await Contact.findByIdAndRemove({ _id: id, owner });
   if (!result) {
     throw HttpError(404, `Contact with a such id=${id} not found`);
   }
@@ -38,6 +52,7 @@ const removeById = async (req, res) => {
 const updateById = async (req, res) => {
   const { error } = schemas.updateSchema.validate(req.body);
   const { id } = req.params;
+  const { _id: owner } = req.user;
 
   if (Object.keys(req.body).length === 0) {
     throw HttpError(404, "Missing fields");
@@ -46,7 +61,10 @@ const updateById = async (req, res) => {
   if (error) {
     throw HttpError(404, `Contact with a such id=${id} not found`);
   }
-  const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
+  const result = await Contact.findOneAndUpdate({ _id: id, owner }, req.body, {
+    new: true,
+  });
+
   if (!result) {
     throw HttpError(404, "Not Found");
   }
@@ -56,6 +74,7 @@ const updateById = async (req, res) => {
 const updateFavorite = async (req, res) => {
   const { error } = schemas.updateFavouriteSchema.validate(req.body);
   const { id } = req.params;
+  const { _id: owner } = req.user;
 
   if (Object.keys(req.body).length === 0) {
     throw HttpError(400, "Missing fields");
@@ -63,7 +82,10 @@ const updateFavorite = async (req, res) => {
   if (error) {
     throw HttpError(404, `Contact with a such id=${id} not found`);
   }
-  const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
+  const result = await Contact.findOneAndUpdate({ _id: id, owner }, req.body, {
+    new: true,
+  });
+
   if (!result) {
     throw HttpError(404, "Not Found");
   }
